% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle_utils.R
\name{compress}
\alias{compress}
\title{A generic function for compressing a data structure.}
\usage{
compress(x, ...)
}
\arguments{
\item{x}{the object to be compressed.}

\item{...}{additional arguments to methods.}
}
\description{
A generic function for compressing a data structure.
}
