% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle-deprecated.R
\name{rle-deprecated}
\alias{rle-deprecated}
\alias{compact.rle}
\title{Deprecated functions from \code{rle}}
\usage{
compact.rle(...)
}
\arguments{
\item{...}{arguments to deprecated functions.}
}
\description{
Deprecated functions from \code{rle}
}
